requirements.txt
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.9
asyncpg
attrs==25.3.0
black==25.1.0
certifi==2025.7.14
click==8.1.8
colorama==0.4.6
dill==0.4.0
fastapi==0.116.1
flake8==7.2.0
frozenlist==1.7.0
greenlet==3.2.3
idna==3.10
isort==6.0.1
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.2
mccabe==0.7.0
multidict==6.6.3
mypy==1.15.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
propcache==0.3.2
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pyflakes==3.3.2
pydantic==2.11.7
pydantic_core==2.33.2
pylint==3.3.7
python-dotenv==1.1.1
ruff==0.11.8
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.1
tomlkit==0.13.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn[standard]
yarl==1.20.1
